pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fenxi28"
version = "1.0.0"
description = "28游戏预测机器人"
dependencies = [
    "loguru>=0.7.0",
    "python-telegram-bot>=13.0",
    "numpy>=1.22.0",
    "requests>=2.28.0",
    "python-dotenv>=0.21.0",
    "pymysql>=1.0.0",
    "schedule>=1.2.0"
]

[tool.setuptools]
packages = ["features"] 